+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-checkout/index.js
@@ -257,7 +257,8 @@ Component.register('swag-paypal-checkout', {
             this.paymentMethods = await this.paymentMethodRepository.search(this.paymentMethodCriteria, Context.api)
                 .then((response) => {
                     return response.filter((paymentMethod) => {
-                        return paymentMethod.formattedHandlerIdentifier !== 'handler_swag_pospayment';
+                        return paymentMethod.formattedHandlerIdentifier !== 'handler_swag_pospayment'
+                            && paymentMethod.formattedHandlerIdentifier !== 'handler_swag_trustlyapmhandler';
                     });
                 });
         },