+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - PPI-685, PPI-701, PPI-725 - Removed auto-hide for Smart Payment Buttons configuration in Administration
 - PPI-731 - Compatibility with Shopware 6.5
 
+# 5.4.6
+- PPI-748 - Disable Trustly for now since PayPal has dropped API support
+- PPI-749 - Fixes an issue, where the fallback button for credit card payments was not processed correctly
+
 # 5.4.5
 - PPI-734 - Fixes an issue, where some payment methods were not displayed in after order process
 - PPI-720, PPI-741, PPI-743 - Fixes an issue, where the total tax amount was calculated incorrectly for net customers